[Step2] 코드리뷰 반영 - Double로 변경, 메서드 체이닝 line break, 익셉션 테스트 메서드명 구체화

+++ src/main/kotlin/step2/Calculator.kt
@@ -2,12 +2,12 @@ package step2
 
 class Calculator {
 
-    fun calculate(input: String): Int {
+    fun calculate(input: String): Double {
         val inputs = input.split(DELIMITER)
-        var result = inputs[0].toInt()
+        var result = inputs[0].toDouble()
         for (i in 1 until inputs.size step 2) {
             val operator = Operator.of(inputs[i].first())
-            result = operator.operate(result, inputs[i + 1].toInt())
+            result = operator.operate(result, inputs[i + 1].toDouble())
         }
 
         return result
